Changes from imageIndex to currentFrame for fence/semaphore DebugMarker/GPUProfiler not working

+++ GraphicsSandbox/Source/Engine/Profiler.cpp
@@ -51,7 +51,7 @@ namespace Profiler
 		if (found != gRangeData.end())
 		{
 			auto& range = found->second;
-			auto dt = (float)range.cpuTimer.elapsedMilliseconds();
+			auto dt = range.cpuTimer.elapsedMilliseconds();
 
 			range.sampleTime += dt;
 			if (found->second.sampleCount == 0)
@@ -104,7 +104,7 @@ namespace Profiler
 			if (v.IsCpuRange())
 				continue;
 
-			float dt = static_cast<float>((queryResult[v.gpuEnd] - queryResult[v.gpuBegin]) * gpuTimestampFrequency * 1e-6);
+			double dt = double((queryResult[v.gpuEnd] - queryResult[v.gpuBegin]) * gpuTimestampFrequency * 1e-6);
 			v.sampleTime += dt;
 			if (v.sampleCount == 0)
 				v.time = dt;
